refact(Categories): reorganize files
