+++ src/reducers/index.ts
@@ -12,8 +12,8 @@ export { default as transactions }         from './transactions';
 export { default as members }              from './members';
 export { default as bankAccounts }         from 'services/bank-accounts/entities-reducer.js';
 export { default as bankAccount }          from 'services/bank-accounts/entity-reducer.js';
-export { default as categories }           from './categories';
-export { default as category }             from './category';
+export { default as categories }           from 'services/categories/entities-reducer.js';
+export { default as category }             from 'services/categories/entity-reducer.js';
 export { default as customers }            from 'services/customers/entities-reducer.js';
 export { default as customer }             from 'services/customers/entity-reducer.js';
 export { default as balances }             from './balances';
