+++ src/services/categories/api.js
@@ -1,4 +1,4 @@
-import { prepareURL, getApi, postApi, putApi, deleteApi } from './_helpers';
+import { prepareURL, getApi, postApi, putApi, deleteApi } from 'api/_helpers.js';
 
 const categoriesURL = (orgId) => prepareURL(`/api/organizations/${orgId}/categories`);
 const categoryURL = (orgId, categoryId) => prepareURL(`/api/organizations/${orgId}/categories/${categoryId}`);
