+++ src/services/categories/types.ts
@@ -1,0 +1,10 @@
+export interface Category {
+  id:       string;
+  name:     string;
+  type:     string;
+}
+
+export interface CategoryParams {
+  name: string;
+  type: string;
+}
